bruh commit

+++ sc.h
@@ -1,4 +1,5 @@
 #include <stdlib.h>
+#include <stdint.h>
 
 #ifndef SC_H
 #define SC_H
@@ -10,8 +11,8 @@ int sc_memoryGet(int addr, int* val);
 int sc_memorySave(char* filename);
 int sc_memoryLoad(char* filename);
 int sc_regInit(void);
-int sc_regSet(int reg, int val);
-int sc_regGet(int reg, int* val);
+int sc_regSet(uint8_t reg, int val);
+int sc_regGet(uint8_t reg, int* val);
 int sc_commandEncode(int cmd, int operand, int* value);
 int sc_commandDecode(int value, int* cmd, int* operand);
